Trace: Provide LogSource, a wrapper for TraceSource
The `LogSource` object wraps a `TraceSource` object, and provides methods for logging that do not rely on enumerations (but bitfields) making logging faster and reduces the need for items to be boxed.

While .NET Core is also compiled, the `TraceSource` object can't read from the `app.config` file, and so always results in the default trace listener, and the test cases would otherwise fail. Future commits should add a different mechanism for tracing.

Issue: DOTNET-362