LogSource: Add support for .NET Core
Support setting the LogSource to a ILoggerFactory or an ILogger that an application can define how a library gets its TraceListener when it also uses this library to instantiate a LogSource.

This allows libraries to use the LogSource and support both .NET Framework, .NET Core, and the application can provide a static instance to avoid having to use dependency injection for console apps that don't need the dependency injection framework.

Applications not using LogSource are not affected, and this is opt-in functionality.

Issue: DOTNET-363