LoggerTraceListener: Don't split lines
A user may have a legitimate reason to write a single event with a message that is internally split into multiple lines with a \n. These should not be split up. One such example is printing a stack trace for an exception, that would result in multiple events, one for each stack frame.

Issue: DOTNET-558