LogSource: Don't overwrite the factory if it is already set
See the `README.md` file for reasons why. A project usually defines its own factory based on the example in section 2.3.1, but test code would also create its own ILoggerFactory. The test code factory would be overwritten as soon as the test code would initialize its logging, which is usually later, thus resulting in incorrect logging in the test code.

Provide the default ability not to overwrite, and instructions on how to use this.

Issue: DOTNET-624